upgrade

+++ app/Http/Controllers/API/invoiceApi.php
@@ -12,6 +12,7 @@ use App\Models\table;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use Maatwebsite\Excel\Facades\Excel;
 
 class invoiceApi extends Controller
@@ -29,12 +30,27 @@ class invoiceApi extends Controller
         $invoiceTotal = 0;
         foreach ($invoices as $invoice) {
             $invoice->date = $invoice->created_at->diffForHumans();
-            $invoiceTotal = $invoiceTotal + $invoice->f_discount;
+            $invoiceTotal = $invoiceTotal + ($invoice->total - ($invoice->total * ($invoice->discount / 100)) + ($invoice->total * ($invoice->tax / 100)));
         }
+
         return [
             'invoices' => $invoices,
             'invoicetotal' => $invoiceTotal
         ];
+    }
+
+    public function notpaid(Request $request)
+    {
+        $positionApi = new positionApi();
+        $check = $positionApi->checkPositionRoute($request->store_id, Auth::id(), 'invoice_add');
+        if ($check) {
+            $store_id = $request->store_id;
+            $store = store::find($request->store_id);
+            $invoices = invoice::where('store_id', $store_id)->where('table_id', 0)->where('paid', NULL)->get();
+            return $invoices;
+        } else {
+            return abort(401);
+        }
     }
 
     public function settings(Request $request)
@@ -82,29 +98,45 @@ class invoiceApi extends Controller
 
     public function deleteinvoice(Request $request)
     {
-        $positionApi = new positionApi();
-        $check = $positionApi->checkPositionRoute($request->store_id, Auth::id(), 'invoice_delete');
-        if ($check) {
-            $invoice = invoice::find($request->invoice_id);
-            if ($invoice) {
-                $store = store::find($invoice->store_id);
-                $historyApi = new historyApi;
-                $des_ar = " تم حذف الفاتورة رقم $invoice->id ";
-                $des_en = " Invoice No. $invoice->id has been deleted ";
-                $history = $historyApi->createHistory($des_ar, $des_en, $store->id, Auth::id());
-                if ($request->table_id && $request->table_id != 0) {
-                    $table = table::find($request->table_id);
-                    if ($table) {
-                        $table->status = 0;
-                        $table->save();
-                    } else {
-                        return abort(404);
+        $this->validate($request, [
+            "password"  => "required"
+        ]);
+        $user = Auth::user();
+        if (Hash::check($request->password, $user->password)) {
+            $positionApi = new positionApi();
+            $check = $positionApi->checkPositionRoute($request->store_id, Auth::id(), 'invoice_delete');
+            if ($check) {
+                $invoice = invoice::find($request->invoice_id);
+                if ($invoice) {
+                    if ($request->table_id) {
+                        if (!$invoice->paid) {
+                            if ($request->table_id != 0) {
+                                $table = table::find($request->table_id);
+                                if ($table) {
+                                    // this Statment to check when the invoice deleted is paid no not
+                                    $table->status = 0;
+                                    $table->save();
+                                } else {
+                                    return abort(404);
+                                }
+                            }
+                        } else {
+                            return abort(404);
+                        }
                     }
+                    $store = store::find($invoice->store_id);
+                    $historyApi = new historyApi;
+                    $des_ar = " تم حذف الفاتورة رقم $invoice->id ";
+                    $des_en = " Invoice No. $invoice->id has been deleted ";
+
+                    $invoice->delete();
+                    $history = $historyApi->createHistory($des_ar, $des_en, $store->id, Auth::id());
+                    return "Deleted successfully";
+                } else {
+                    return abort(404);
                 }
-                $invoice->delete();
-                return "Deleted successfully";
             } else {
-                return abort(404);
+                return abort(401);
             }
         } else {
             return abort(401);
@@ -124,29 +156,33 @@ class invoiceApi extends Controller
         $check = $positionApi->checkPositionRoute($request->store_id, Auth::id(), 'invoice_add');
         if ($check) {
             $invoice = invoice::find($request->invoice_id);
-            $store = store::find($invoice->store_id);
-            if ($request->table_id != 0) {
-                $table = table::find($request->table_id);
-                if ($invoice->store_id == $table->store_id) {
-                    if ($request->paidamount >= $invoice->f_discount) {
-                        $table->status = 0;
-                        $table->save();
-                        $invoice->paid = $request->f_discount;
-                        $invoice->save();
-                        return 'Invoice Completed';
+            if ($invoice) {
+                $store = store::find($invoice->store_id);
+                $total = $invoice->total - ($invoice->total * ($invoice->discount / 100)) + ($invoice->total * ($invoice->tax / 100));
+                if ($request->table_id != 0) {
+                    $table = table::find($request->table_id);
+                    if ($invoice->store_id == $table->store_id) {
+                        if ($request->paidamount >= $total) {
+                            $table->status = 0;
+                            $table->save();
+                            $invoice->paid = $total;
+                            return $invoice->save();
+                        } else {
+                            return abort(419);
+                        }
                     } else {
                         return abort(419);
                     }
                 } else {
-                    return abort(419);
+                    if ($request->paidamount >= $invoice->f_discount) {
+                        $invoice->paid = $total;
+                        return $invoice->save();
+                    } else {
+                        return abort(419);
+                    }
                 }
             } else {
-                if ($request->paidamount >= $invoice->f_discount) {
-                    $invoice->paid = $invoice->f_discount + ($invoice->f_discount * ($invoice->tax  / 100));
-                    return $invoice->save();
-                } else {
-                    return abort(419);
-                }
+                return abort(401);
             }
         } else {
             return abort(401);
